+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+---
+
+## [0.7.8] - 2025-10-07
+
+ - [Jobs pagination](https://app.asana.com/1/355454733208534/project/1210341690761798/task/1211792018941738)
+   - Hide navigation dots when a job is open
+
+ - [Collaborators](https://app.asana.com/1/355454733208534/project/1210341690761798/task/1211792018941748)
+     - Hide season buttons on mobile
+     - New block collaborator functionality
+
+
 ## [0.7.7] - 2025-10-06
 
 ### Added